LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-030-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: IntersectionObserver

Harness Error (FAIL), message = ReferenceError: Can't find variable: IntersectionObserver

TIMEOUT IntersectionObserver interactions Test timed out

Harness Error (FAIL), message = ReferenceError: Can't find variable: IntersectionObserver

TIMEOUT IntersectionObserver interactions Test timed out
